Rename constants without "copy" suffix
