+++ ui/src/admin/actions/chronograf.js
@@ -18,7 +18,7 @@ import {
 
 import {publishNotification} from 'shared/actions/notifications'
 import {errorThrown} from 'shared/actions/errors'
-import {chronografUserNotifications} from 'shared/copy/notificationsCopy'
+import {chronografUserNotifications} from 'shared/copy/notifications'
 
 import {REVERT_STATE_DELAY} from 'shared/constants'
