+++ ui/src/admin/components/chronograf/AllUsersTableRowNew.js
@@ -6,7 +6,7 @@ import {bindActionCreators} from 'redux'
 import {publishNotification as publishNotificationAction} from 'shared/actions/notifications'
 import Dropdown from 'shared/components/Dropdown'
 
-import {chronografUserNotifications} from 'shared/copy/notificationsCopy'
+import {chronografUserNotifications} from 'shared/copy/notifications'
 import {ALL_USERS_TABLE} from 'src/admin/constants/chronografTableSizing'
 const {
   colOrganizations,
