+++ ui/src/shared/dispatchers/index.js
@@ -1,6 +1,6 @@
 import {publishNotification} from 'shared/actions/notifications'
 import {delayEnablePresentationMode} from 'shared/actions/app'
-import {enterPresentationModeNotification} from 'shared/copy/notificationsCopy'
+import {enterPresentationModeNotification} from 'shared/copy/notifications'
 
 export const presentationButtonDispatcher = dispatch => () => {
   dispatch(delayEnablePresentationMode())
